feat(dashboard): show recent events for all node types
Lift event fetching from IrrigationControl and CurtainControl into
NodeDetail so every node type (sensor, irrigation, greenhouse, etc.)
displays the shared RecentEvents timeline component.
